WIP: Remove legacy metadata from command implementations
