+++ Sources/D2Commands/misc/PingCommand.swift
@@ -2,11 +2,14 @@ import SwiftDiscord
 import D2Permissions
 
 public class PingCommand: Command {
-	public let description = "Replies with 'Pong!'"
+	public let info = CommandInfo(
+		category: .misc,
+		shortDescription: "Replies with 'Pong!'",
+		longDescription: "Outputs 'Pong!'",
+		requiredPermissionLevel: .basic
+	)
 	public let inputValueType: RichValueType = .none
 	public let outputValueType: RichValueType = .text
-	public let sourceFile: String = #file
-	public let requiredPermissionLevel = PermissionLevel.basic
 	
 	public init() {}
